+++ Example/test/fetch.spec.js
@@ -103,7 +103,7 @@ describe("Fetch AuthHttpRequest class tests", function() {
     });
 
     it("testing with fetch api methods without config", async function() {
-        AuthHttpRequestFetch.init(`${BASE_URL}/refresh`);
+        AuthHttpRequestFetch.init(`${BASE_URL}/refresh`, false);
 
         let getResponse = await AuthHttpRequestFetch.get(`${BASE_URL}/testing`);
         let postResponse = await AuthHttpRequestFetch.post(`${BASE_URL}/testing`);
@@ -169,7 +169,7 @@ describe("Fetch AuthHttpRequest class tests", function() {
     });
 
     it("testing with fetch api methods that doesn't exists", async function() {
-        AuthHttpRequestFetch.init(`${BASE_URL}/refresh`);
+        AuthHttpRequestFetch.init(`${BASE_URL}/refresh`, false);
 
         let getResponse = await AuthHttpRequestFetch.get(`${BASE_URL}/fail`);
         let postResponse = await AuthHttpRequestFetch.post(`${BASE_URL}/fail`);
@@ -303,7 +303,7 @@ describe("Fetch AuthHttpRequest class tests", function() {
             jest.setTimeout(15000);
             await startST(5);
 
-            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`, true);
+            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`);
             let userId = "testing-supertokens-react-native";
 
             //send loing request
@@ -330,7 +330,7 @@ describe("Fetch AuthHttpRequest class tests", function() {
             jest.setTimeout(15000);
             await startST(5);
 
-            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`, true);
+            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`);
             let userId = "testing-supertokens-react-native";
 
             // send api request to login
@@ -370,7 +370,7 @@ describe("Fetch AuthHttpRequest class tests", function() {
             jest.setTimeout(15000);
             await startST(5);
 
-            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`, true);
+            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`);
             let userId = "testing-supertokens-react-native";
 
             // send api request to login
@@ -422,7 +422,7 @@ describe("Fetch AuthHttpRequest class tests", function() {
             jest.setTimeout(15000);
             await startST(3, false);
 
-            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`, true);
+            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`);
             let userId = "testing-supertokens-react-native";
 
             // send api request to login
@@ -486,7 +486,7 @@ describe("Fetch AuthHttpRequest class tests", function() {
             jest.setTimeout(15000);
             await startST();
 
-            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`, true);
+            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`);
 
             let val = await global.fetch(`${BASE_URL}/testError`);
             assertEqual(await val.text(), "test error message");
@@ -522,8 +522,8 @@ describe("Fetch AuthHttpRequest class tests", function() {
             jest.setTimeout(15000);
             await startST();
 
-            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`, true);
-            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`, true);
+            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`);
+            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`);
             let userId = "testing-supertokens-react-native";
 
             let loginResponse = await global.fetch(`${BASE_URL}/login`, {
@@ -576,7 +576,7 @@ describe("Fetch AuthHttpRequest class tests", function() {
             jest.setTimeout(15000);
             await startST(3);
 
-            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`, true);
+            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`);
             let userId = "testing-supertokens-react-native";
 
             // send api request to login
@@ -614,7 +614,7 @@ describe("Fetch AuthHttpRequest class tests", function() {
             jest.setTimeout(15000);
             await startST(5);
 
-            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`, true);
+            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`);
             let userId = "testing-supertokens-react-native";
 
             let loginResponse = await global.fetch(`${BASE_URL}/login`, {
@@ -657,7 +657,7 @@ describe("Fetch AuthHttpRequest class tests", function() {
             jest.setTimeout(15000);
             await startST(5);
 
-            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`, true);
+            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`);
             let userId = "testing-supertokens-react-native";
 
             // send api request to login
@@ -692,7 +692,7 @@ describe("Fetch AuthHttpRequest class tests", function() {
         try {
             jest.setTimeout(15000);
             await startST(5);
-            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`, true);
+            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`);
             let userId = "testing-supertokens-react-native";
 
             // this is technically not doing interception, but it is equavalent to doing it since the inteceptor just calls the function below.
@@ -731,7 +731,7 @@ describe("Fetch AuthHttpRequest class tests", function() {
         try {
             jest.setTimeout(15000);
             await startST(5);
-            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`, true);
+            AuthHttpRequestFetch.init(`${BASE_URL}/refresh`);
             let userId = "testing-supertokens-react-native";
 
             let myFetch = async (url, config) => {